yumlabs/requirements/venv.txt
absl-py==2.2.2
accelerate==1.6.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
basicsr==1.4.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.2
controlnet-aux==0.0.9
cycler==0.12.1
diffusers==0.32.2
einops==0.8.1
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
loguru==0.7.3
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.3.1
peft==0.15.1
pillow==11.1.0
platformdirs==4.3.7
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.3
tb-nightly==2.20.0a20250417
tensorboard-data-server==0.7.2
tifffile==2025.3.30
timm==0.6.7
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.182
ultralytics-thop==2.0.16
urllib3==2.3.0
Werkzeug==3.1.3
yapf==0.43.0
zipp==3.21.0
